backend/requirements.txt
# ShieldIaC Backend — Production Dependencies

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# HTTP
httpx==0.27.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35

# Redis
redis[hiredis]==5.1.0

# YAML parsing (for Kubernetes/CloudFormation scanning)
pyyaml==6.0.2

# OpenAI (AI fix suggestions)
openai==1.50.0

# PDF report generation
reportlab==4.2.2

# Stripe billing
stripe==10.10.0

# Security
cryptography==43.0.0
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
